feat(batch): Improve data sync and monitoring capabilities
- Add docker-compose.yml for standalone service execution
- Enhance data synchronization with better error handling
- Improve IBKR connection monitoring and recovery
- Update database configurations for better performance
- Add comprehensive README documentation
- Remove outdated test files

Changes include:
- Enhanced CollectDataService with retry logic
- Improved DataSyncService with batch processing
- Better connection pool management in JdbcDataSourceRouter
- Updated application.yaml with production-ready settings

+++ trade_batch/src/main/java/com/trade/batch/service/DataSyncService.java
@@ -450,6 +450,6 @@ public class DataSyncService {
     }
 
     private void incrementSyncCount(String table, int count) {
-        syncStatistics.merge(table, count, (old, inc) -> ((Integer) old) + inc);
+        syncStatistics.merge(table, count, (old, inc) -> (Integer) old + (Integer) inc);
     }
 }
